refactor(ui): extract shared useOverflowBadges hook
The single-line overflow measurement (hidden measurement layer +
ResizeObserver + indicator-width reservation) was duplicated between
MultiSelectBadges (case custom fields) and SelectedToolsHeader (chat
composer). Extract it into a shared useOverflowBadges hook so both
surfaces share one implementation and cannot drift.

Addresses the cubic review note on #2954.
