+++ frontend/src/hooks/index.ts
@@ -4,3 +4,4 @@ export * from "./use-agent-channels"
 export * from "./use-agent-presets"
 export * from "./use-debounce"
 export * from "./use-entitlements"
+export * from "./use-overflow-badges"
